remote_tools_folders/llm-call-audienceEdit-think-gemini/requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
openai>=1.16.0
Pillow>=9.0.0 